fix(config-resolver): add options for utility functions
