--- packages/config-resolver/src/endpointsConfig/configurations.ts
@@ -1,32 +1,0 @@
-import { Endpoint, Provider, RegionInfoProvider, UrlParser } from "@aws-sdk/types";
-
-export interface EndpointsInputConfig {
-  /**
-   * The fully qualified endpoint of the webservice. This is only required when using a custom endpoint (for example, when using a local version of S3).
-   */
-  endpoint?: string | Endpoint | Provider<Endpoint>;
-
-  /**
-   * Whether TLS is enabled for requests.
-   */
-  tls?: boolean;
-}
-
-export interface PreviouslyResolved {
-  regionInfoProvider: RegionInfoProvider;
-  urlParser: UrlParser;
-  region: Provider<string>;
-}
-
-export interface EndpointsResolvedConfig extends Required<EndpointsInputConfig> {
-  /**
-   * Resolved value for input {@link EndpointsResolvedConfig.endpoint}
-   */
-  endpoint: Provider<Endpoint>;
-
-  /**
-   * Whether the endpoint is specified by caller.
-   * @internal
-   */
-  isCustomEndpoint: boolean;
-}
