+++ packages/config-resolver/src/endpointsConfig/index.ts
@@ -1,2 +1,1 @@
-export * from "./configurations";
 export * from "./resolveEndpointsConfig";
